Remove classes specific to Open AI implementation and remove Open AI dependencies. Prepare project to use only Azure Open AI services.
