--- src/main/java/com/example/rag/qa/QAController.java
@@ -1,33 +1,0 @@
-package com.example.rag.qa;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Profile;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.LinkedHashMap;
-import java.util.Map;
-
-@RestController
-@RequestMapping("/qa")
-@Profile("openai")
-public class QAController {
-
-    private final QAService qaService;
-
-    @Autowired
-    public QAController(QAService qaService) {
-        this.qaService = qaService;
-    }
-
-    @GetMapping
-    public Map completion(@RequestParam(value = "question") String question) {
-        String answer = this.qaService.generate(question);
-        Map map = new LinkedHashMap();
-        map.put("question", question);
-        map.put("answer", answer);
-        return map;
-    }
-}
